AWS-Rekognition/requirements.txt
boto3==1.17.77
botocore==1.20.77
jmespath==0.10.0
numpy==1.20.3
opencv-contrib-python==4.5.2.52
Pillow==8.2.0
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
